Move netcdf file handler class to separate module from VIIRS L1B reader
Also prepare generic reader for handling other dimensions besides 2D.
