+++ satpy/readers/viirs_l1b.py
@@ -25,73 +25,24 @@
 """Interface to VIIRS L1B format
 
 """
-import os.path
-from datetime import datetime, timedelta
-import numpy as np
 import logging
+from datetime import datetime, timedelta
 
-from satpy.readers import ConfigBasedReader, MultiFileReader, FileKey, GenericFileReader
-import netCDF4
-import six
+import numpy as np
+
+from satpy.readers import ConfigBasedReader, GenericFileReader
+from satpy.readers.netcdf_utils import NetCDF4FileWrapper
 
 NO_DATE = datetime(1958, 1, 1)
 EPSILON_TIME = timedelta(days=2)
 LOG = logging.getLogger(__name__)
 
 
-class NetCDF4MetaData(object):
-    """Small class for inspecting a NetCDF4 file and retrieve its metadata/header data.
-    """
-    def __init__(self, filename, auto_maskandscale=False, **kwargs):
-        self.metadata = {}
-        self.filename = filename
-        if not os.path.exists(filename):
-            raise IOError("File %s does not exist!" % filename)
-        file_handle = netCDF4.Dataset(self.filename, 'r')
-
-        self.auto_maskandscale= auto_maskandscale
-        if hasattr(file_handle, "set_auto_maskandscale"):
-            file_handle.set_auto_maskandscale(auto_maskandscale)
-
-        self.collect_metadata("", file_handle)
-        file_handle.close()
-
-    def _collect_attrs(self, name, obj):
-        for key in obj.ncattrs():
-            value = getattr(obj, key)
-            value = np.squeeze(value)
-            if issubclass(value.dtype.type, str) or np.issubdtype(value.dtype, np.character):
-                self.metadata["%s/attr/%s" % (name, key)] = str(value)
-            else:
-                self.metadata["%s/attr/%s" % (name, key)] = value
-
-    def collect_metadata(self, name, obj):
-        # Look through each subgroup
-        base_name = name + "/" if name else ""
-        for group_name, group_obj in obj.groups.items():
-            self.collect_metadata(base_name + group_name, group_obj)
-        for var_name, var_obj in obj.variables.items():
-            var_name = base_name + var_name
-            self.metadata[var_name] = var_obj
-            self.metadata[var_name + "/shape"] = var_obj.shape
-            self._collect_attrs(var_name, var_obj)
-        self._collect_attrs(name, obj)
-
-    def __getitem__(self, key):
-        val = self.metadata[key]
-        if isinstance(val, netCDF4.Variable):
-            # these datasets are closed and inaccessible when the file is closed, need to reopen
-            v = netCDF4.Dataset(self.filename, 'r')
-            val = v[key]
-            val.set_auto_maskandscale(self.auto_maskandscale)
-        return val
-
-
 class L1BFileReader(GenericFileReader):
-    """VIIRS HDF5 File Reader
+    """VIIRS L1B File Reader
     """
     def create_file_handle(self, filename, **kwargs):
-        handle = NetCDF4MetaData(filename, **kwargs)
+        handle = NetCDF4FileWrapper(filename, **kwargs)
         return handle.filename, handle
 
     def __getitem__(self, item):
